refactor: move the interfaces to a separate folder
