+++ react-components/src/components/cardList/CardsList.tsx
@@ -1,30 +1,10 @@
 import React, { useState } from 'react';
 import { ProductCard } from './ProductCard';
 import { Modal } from '../modal/Modal';
+import Card from '../../models/ICard'
 
 type CardsListProps = {
   characters: Card[];
-};
-
-export type Card = {
-  id: number;
-  name: string;
-  status: string;
-  species: string;
-  type: string;
-  gender: string;
-  origin: {
-    name: string;
-    url: string;
-  };
-  location: {
-    name: string;
-    url: string;
-  };
-  image: string;
-  episode: string[];
-  url: string;
-  created: string;
 };
 
 function CardsList(props: CardsListProps) {
