+++ react-components/src/components/cardList/ProductCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Card } from '../cardList/CardsList';
+import Card  from '../../models/ICard';
 
 type ProductCardProps = {
   click: (content: Card) => void;
