+++ react-components/src/models/ICard.ts
@@ -1,0 +1,20 @@
+export default interface Card {
+    id: number;
+    name: string;
+    status: string;
+    species: string;
+    type: string;
+    gender: string;
+    origin: {
+      name: string;
+      url: string;
+    };
+    location: {
+      name: string;
+      url: string;
+    };
+    image: string;
+    episode: string[];
+    url: string;
+    created: string;
+};
