+++ react-components/src/models/IUser.ts
@@ -1,0 +1,7 @@
+export default interface User {
+    name?: string;
+    date?: string;
+    gender?: string;
+    img?: string;
+    country?: string;
+};
